Extract formatValue to FormatUtils
