+++ client/trino-cli/src/main/java/io/trino/cli/CsvPrinter.java
@@ -15,17 +15,12 @@ package io.trino.cli;
 
 import au.com.bytecode.opencsv.CSVWriter;
 import com.google.common.collect.ImmutableList;
-import io.trino.client.Row;
 
 import java.io.IOException;
 import java.io.Writer;
 import java.util.List;
-import java.util.Map;
 
-import static io.trino.cli.AlignedTablePrinter.formatHexDump;
-import static io.trino.cli.AlignedTablePrinter.formatList;
-import static io.trino.cli.AlignedTablePrinter.formatMap;
-import static io.trino.cli.AlignedTablePrinter.formatRow;
+import static io.trino.cli.FormatUtils.formatValue;
 import static java.util.Objects.requireNonNull;
 
 public class CsvPrinter
@@ -114,33 +109,8 @@ public class CsvPrinter
             array = new String[rowSize];
         }
         for (int i = 0; i < rowSize; i++) {
-            array[i] = formatValue(values.get(i));
+            array[i] = formatValue(values.get(i), "", -1);
         }
         return array;
-    }
-
-    static String formatValue(Object o)
-    {
-        if (o == null) {
-            return "";
-        }
-
-        if (o instanceof Map) {
-            return formatMap((Map<?, ?>) o);
-        }
-
-        if (o instanceof List) {
-            return formatList((List<?>) o);
-        }
-
-        if (o instanceof Row) {
-            return formatRow((Row) o);
-        }
-
-        if (o instanceof byte[]) {
-            return formatHexDump((byte[]) o);
-        }
-
-        return o.toString();
     }
 }
