+++ client/trino-cli/src/main/java/io/trino/cli/JsonPrinter.java
@@ -24,7 +24,7 @@ import java.io.IOException;
 import java.io.Writer;
 import java.util.List;
 
-import static io.trino.cli.AlignedTablePrinter.formatHexDump;
+import static io.trino.cli.FormatUtils.formatHexDump;
 import static java.util.Objects.requireNonNull;
 
 public class JsonPrinter
